refactor: abstract logic of listResults into a custom hook
